requirements.txt
pandas
openpyxl
plotly.express
streamlit
google-auth
gspread
xlrd

